Implements a new Multi-Day Calendar View Event Grid
This commit replaces the existing Multi-day Calendar View that previously
used absolute positioning to layout events, replacing it to a more "Native"
approach using CSS Grid and moving the responsibility to deal with sizes
to the browser itself.

It is also expected to render events faster, since it does less floating point
math operations and to have less layout shifting when creating, editing and
deleting events.

Co-authored-by: and <and@tutao.de>