[ios] Fix resetting badge on app startup
`applicationWillEnterForeground` is not getting called on app
startup, only on transition from background to foreground state.
This means the counter would only get reset when the app would
come foreground from background state.

We did not notice it as notifications are also reset via
NativePushFacade.

We fixed it by using `applicationDidBecomeActive` instead. It it a
different lifecycle event but it also works, perhaps even better.

Close #9922

Co-authored-by: hrb-hub <hrb-hub@users.noreply.github.com>